+++ service/src/main/java/com/ligg/service/common/impl/ProjectServiceImpl.java
@@ -87,13 +87,13 @@ public class ProjectServiceImpl extends ServiceImpl<ProjectMapper, ProjectEntity
      * 添加项目
      */
     @Override
-    public void saveProject(String projectName, Double projectPrice) {
+    public void saveProject(String projectName, Float projectPrice) {
 
-            ProjectEntity projectEntity = new ProjectEntity();
-            projectEntity.setProjectName(projectName);
-            projectEntity.setProjectPrice(projectPrice);
-            projectEntity.setProjectCreatedAt(LocalDateTime.now());
-            projectMapper.insert(projectEntity);
+        ProjectEntity projectEntity = new ProjectEntity();
+        projectEntity.setProjectName(projectName);
+        projectEntity.setProjectPrice(projectPrice);
+        projectEntity.setProjectCreatedAt(LocalDateTime.now());
+        projectMapper.insert(projectEntity);
     }
 
     /**
@@ -104,6 +104,7 @@ public class ProjectServiceImpl extends ServiceImpl<ProjectMapper, ProjectEntity
         return projectMapper.selectOne(new LambdaQueryWrapper<ProjectEntity>()
                 .eq(ProjectEntity::getProjectName, projectName)) != null;
     }
+
     /**
      * 编辑项目
      */
@@ -115,6 +116,4 @@ public class ProjectServiceImpl extends ServiceImpl<ProjectMapper, ProjectEntity
                 .set(ProjectEntity::getProjectName, projectName)
                 .set(ProjectEntity::getProjectUpdateAt, LocalDateTime.now()));
     }
-
-
 }
